begin of modification Update method in db context

+++ DepartmentWebApi/code/Interfaces/IDepartmentRepository.cs
@@ -5,9 +5,9 @@ namespace Interfaces{
 	public interface IDepartmentRepository
 	{
 		IEnumerable<Department> GetAll();
-		Department Get(int id);
+		Department Get(long id);
 		int Insert(Department department);
 		int Update(Department department);
-		int Delete(int id);
+		int Delete(long id);
 	}
 }
